Add Ruby 1.9.3 back to Travis test list + official 2.3
I hate to flip flop on this, but I decided to patch json_schema to still
support 1.9.3 because it only necessitated a change in a single line. As
such, I'm bringing it back to Committee just to keep the version on life
support a bit longer (the real reason is I want to use it in
stripe-ruby's test suite and we're currently supporting versions of Ruby
well past the expiration dates right now, including 1.9.3).

Also move to 2.3.0 from 2.3.0-preview2 now that we have an official
release.