Require driver for "committee-stub" executable
This tweaks the behavior of `committee-stub` a bit so that a driver is
now required to start it (i.e. either `hyper_schema` or `open_api_2`).

The reason to do this is that OpenAPI 2.0 will generally parse as a
valid hyper-schema, just one with no links and which doesn't do anything
useful. This leads to very confusing gotcha behavior in that you boot
the stub, nothing works, and it's not obvious what's going on.

This breaks the CLI API, but it's worth it. Fewer users should be using
this program and it's still semver-consistent with the coming major
version bump.