Deprecate method.rb methods
Now method.rb depends on JSON Hyper-Schema structre.
So we need change for OpenAPI3.

In OpenAPI3 branch we refactoring abstruct schema object, so we don't need know detail structre.
We can change method.rb like this code.
https://github.com/interagent/committee/blob/485a863ecade369e942528d1ab769ddf829d57cb/lib/committee/test/methods.rb

And in 3.0, we should use Schema object.
https://github.com/interagent/committee#set-committeedriversschema-object-for-middleware
The test method should use Schema object too so we deprecated other methods and drop in 3.0.