change validate_errors to validate_success_only
The validate_erros option validate 1xx and 3xx status code but it's not error status code.
So we should deprecated and change option name.

This change is deprecated only and we'll change new option when 3.x released.

(We can all status code schema in OpenAPI3 so it will be more large problem)