Perform a strict check on properties
We need to throw a validation error if property being send is
not defined. If additionalProperties are defined, we skip the
validation for now, with TODO added.

In a case of composing schema using allOf, the property needs
to be defined at least in 1 of te allOf items, as stated in:

https://swagger.io/docs/specification/data-models/oneof-anyof-allof-not/#allof