Add new app to test examples from the readme
In this commit, we are adding a new Sinatra app "heroku_api" to the
examples folder.

This new app provides the endpoints documented in the readme of the
project, so we can use it to check how Committee is handling it.

The new app was implemented following a very simplistic approach,
so it has the minimum to work as a testing tool for the exemples
provided in the documentation of the project.